requirements.txt
absl-py==2.3.1
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.0
filetype==1.2.0
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.10.0
idna==3.7
jax==0.8.0
jaxlib==0.8.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mediapipe==0.10.14
ml_dtypes==0.5.3
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
packaging==25.0
pi_heif==1.1.1
pillow==12.0.0
pillow-avif-plugin==1.5.2
polars==1.35.2
polars-runtime-32==1.35.2
protobuf==4.25.8
psutil==7.1.3
pycparser==2.23
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
roboflow==1.2.11
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
sounddevice==0.5.3
sympy==1.14.0
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
typing_extensions==4.15.0
ultralytics==8.3.228
ultralytics-thop==2.0.18
urllib3==2.5.0
wheel==0.45.1
